requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
google-cloud-secret-manager
google-genai
